requirements.txt
faiss-cpu==1.7.4
Flask==2.3.2
google-api-python-client==2.93.0
huggingface-hub==0.16.4
inspiredco==0.0.2
keras==2.13.1
langchain==0.0.153
openai==0.27.8
pdfplumber==0.9.0
tensorflow==2.13.0
tiktoken==0.4.0
tokenizers==0.13.3
transformers==4.30.2
python-dotenv==1.0.0